requirements.txt
beautifulsoup4==4.6.0
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.2
django-modelcluster==4.3
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.9.2
draftjs-exporter==2.1.5
gunicorn==19.7.1
html5lib==1.0.1
idna==2.8
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
requests==2.21.0
six==1.12.0
Unidecode==1.0.23
urllib3==1.24.2
wagtail==2.4
webencodings==0.5.1
whitenoise==3.3.1
Willow==1.1
